Use buble dist, babel-register for tests (#12)
* Play with buble build for dist, babel 6 register for tests.

* Add node 6 to travis. Stop using includes()

* Compile to es5 + imports, use babel-register for tests, rollup for dist

This experiment uses buble to build the src and tests to build/src and
build/test, without transforming es2015 modules.  Then, it uses
babel-register with *only* the es2015 to commonjs module transform to
run the unit tests.  Finally, it uses rollup, with no additional
plugins to build the dist files.

Code coverage is completely broken in this setup, but it works well
otherwise, and avoids the problem of testing babel-generated code
while publishing buble-generated code.

* Update buble and un-quote methods

* Remove leftover use strict